// test/protractor/pageobjects/enum/enum.ts
export enum gender{
  MALE = 'male',
  FEMALE = 'female'
}
export enum prefix{
  NO_PREFIX = ''
}
export enum houseNumberAddition{
  NO_HOUSENUMBER_ADDITION = ''
}
export enum phoneNumber{
  NO_PHONENUMBER = ''
}
export enum specificIdentification{
  PASSPORT = 'passport',
  ID_CARD = 'Id card',
  DRIVER_LICENSE = 'Driver license',
  SOMETHING_ELSE =  'Something else',
  NO_NUMBER = ''
}

export enum personaName{
  RONALDO = 'ronaldo',
  MESSI = 'messi',
  SALAH = 'salah',
  NEYMAR = 'neymar',
  POGBA = 'pogba'
}

export enum yesNo {
  YES = 'yes',
  NO = 'no',
  EMPTY = ''
}